branch3 refactor
